# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Point of Sales System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List all suppliers
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Customers
...
  /sales-returns:
    get:
      operationId: listSalesReturns
      tags: [salesReturns]
      summary: List all sales returns
      parameters:
        - name: customerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: from
          in: query
          schema:
            type: ["string", "null"]
        - name: to
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of sales returns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesReturnWithDetails"
    post:
      operationId: createSalesReturn
      tags: [salesReturns]
      summary: Create a sales return
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSalesReturnBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesReturnWithDetails"
  /sales-returns/{id}:
    get:
      operationId: getSalesReturn
      tags: [salesReturns]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sales return with details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesReturnWithDetails"
        "404":
          description: Not found

  # Stock movements
...
components:
  schemas:
...
    Supplier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        contactPerson:
          type: ["string", "null"]
        contactNo:
          type: ["string", "null"]
        creditLimit:
          type: number
        createdAt:
          type: string
      required:
        - id
        - name
        - creditLimit
        - createdAt
    CreateSupplierBody:
      type: object
      properties:
        name:
          type: string
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        contactPerson:
          type: ["string", "null"]
        contactNo:
          type: ["string", "null"]
        creditLimit:
          type: number
      required:
        - name
        - creditLimit
...
    SalesReturnDetail:
      type: object
      properties:
        id:
          type: integer
        salesReturnId:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productCode:
          type: string
        salesPrice:
          type: number
        discount:
          type: number
        qty:
          type: number
      required:
        - id
        - salesReturnId
        - productId
        - productName
        - productCode
        - salesPrice
        - discount
        - qty
    SalesReturn:
      type: object
      properties:
        id:
          type: integer
        returnDate:
          type: string
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        originalBillNo:
          type: ["string", "null"]
        totalAmount:
          type: number
        createdAt:
          type: string
      required:
        - id
        - returnDate
        - totalAmount
        - createdAt
    SalesReturnWithDetails:
      allOf:
        - $ref: "#/components/schemas/SalesReturn"
        - type: object
          properties:
            details:
              type: array
              items:
                $ref: "#/components/schemas/SalesReturnDetail"
          required:
            - details
    CreateSalesReturnDetailItem:
      type: object
      properties:
        productId:
          type: integer
        salesPrice:
          type: number
        discount:
          type: number
        qty:
          type: number
      required:
        - productId
        - salesPrice
        - qty
    CreateSalesReturnBody:
      type: object
      properties:
        returnDate:
          type: string
        customerId:
          type: ["integer", "null"]
        originalBillNo:
          type: ["string", "null"]
        details:
          type: array
          items:
            $ref: "#/components/schemas/CreateSalesReturnDetailItem"
      required:
        - returnDate
        - details
...