# lib/api-zod/src/generated/types/createSalesReturnDetailItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Point of Sales System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSalesReturnDetailItem {
  productId: number;
  salesPrice: number;
  discount?: number;
  qty: number;
}